Resource representation proposal
This is a little odd for a proposal because we've already been working on it
for a while, but it seems like a good idea to put this high level overview here
with the other proposals as a breadcrumb. The text here is a cleaned up version
of microsoft/hlsl-specs#223 with a few updates based on the changes that have
happened since.