Update site publishing to fix title and edit page link (#335)
This updates the hugo configuration so that the "Edit on github" links
work.

`README.md` is now directly mounted as `_index.md`, with the most
noticeable difference being that the title on the home page is now "HLSL
Working Group" rather than "HLSL Working Group - HLSL Working Group".

Also, a small tweak to hugo.toml so that the `ignoreLogs` value is
top-level.