Mapping of GetDimensions overload to builtin functions and LLVM intrinsics
This proposal outlines how to map the many GetDimensions overloads to Clang builtin functions and LLVM intrinsics for both buffer resources and for textures. There are 54 cases of GetDimensions we need to support. We should be able to reduce this to a much lower number of builtin function that will get translated to one or more LLVM intrinsics.

Closes #346