Mapping of GetDimensions overload to builtin functions and LLVM intrinsics (#349)
This proposal describes how to map the many `GetDimensions` overloads for buffer and texture resources to Clang built-in functions and LLVM intrinsics. There are 54 distinct `GetDimensions` cases to support and we can significantly reduce the number of required built-in functions by grouping overloads with similar argument patterns. Each built-in function can then be translated to one or more LLVM intrinsics depending on target platform.

Closes #346